Repair / implement folders endpoint
