+++ src/core/Directus/Services/FilesServices.php
@@ -211,75 +211,6 @@ class FilesServices extends AbstractService
         return $this->getItemsAndSetResponseCacheTags($tableGateway, $params);
     }
 
-    public function createFolder(array $data, array $params = [])
-    {
-        $collection = 'directus_folders';
-        $this->enforceCreatePermissions($collection, $data, $params);
-        $this->validatePayload($collection, null, $data, $params);
-
-        $foldersTableGateway = $this->createTableGateway($collection);
-
-        $newFolder = $foldersTableGateway->createRecord($data, $this->getCRUDParams($params));
-
-        return $foldersTableGateway->wrapData(
-            $newFolder->toArray(),
-            true,
-            ArrayUtils::get($params, 'meta')
-        );
-    }
-
-    public function findFolder($id, array $params = [])
-    {
-        $foldersTableGateway = $this->createTableGateway('directus_folders');
-        $params['id'] = $id;
-
-        return $this->getItemsAndSetResponseCacheTags($foldersTableGateway, $params);
-    }
-
-    public function findFolderByIds($id, array $params = [])
-    {
-        $foldersTableGateway = $this->createTableGateway('directus_folders');
-
-        return $this->getItemsByIdsAndSetResponseCacheTags($foldersTableGateway, $id, $params);
-    }
-
-    public function updateFolder($id, array $data, array $params = [])
-    {
-        $collectionName = 'directus_folders';
-        $this->enforceUpdatePermissions($collectionName, $data, $params);
-        $this->checkItemExists($collectionName, $id);
-
-        $foldersTableGateway = $this->createTableGateway('directus_folders');
-        $group = $foldersTableGateway->updateRecord($id, $data, $this->getCRUDParams($params));
-
-        return $foldersTableGateway->wrapData(
-            $group->toArray(),
-            true,
-            ArrayUtils::get($params, 'meta')
-        );
-    }
-
-    public function findAllFolders(array $params = [])
-    {
-        $foldersTableGateway = $this->createTableGateway('directus_folders');
-
-        return $this->getItemsAndSetResponseCacheTags($foldersTableGateway, $params);
-    }
-
-    public function deleteFolder($id, array $params = [])
-    {
-        $this->enforcePermissions('directus_folders', [], $params);
-
-        $foldersTableGateway = $this->createTableGateway('directus_folders');
-        // NOTE: check if item exists
-        // TODO: As noted in other places make a light function to check for it
-        $this->getItemsAndSetResponseCacheTags($foldersTableGateway, [
-            'id' => $id
-        ]);
-
-        return $foldersTableGateway->deleteRecord($id, $this->getCRUDParams($params));
-    }
-
     /**
      * @param array $items
      * @param array $params
